+++ store/src/main/java/org/apache/rocketmq/store/ha/protocol/HandshakeSlave.java
@@ -8,7 +8,7 @@ public class HandshakeSlave {
 
     private String brokerName;
 
-    private int brokerId;
+    private long brokerId;
 
     private int brokerAppVersion;
 
@@ -32,11 +32,11 @@ public class HandshakeSlave {
         this.brokerName = brokerName;
     }
 
-    public int getBrokerId() {
+    public long getBrokerId() {
         return brokerId;
     }
 
-    public void setBrokerId(int brokerId) {
+    public void setBrokerId(long brokerId) {
         this.brokerId = brokerId;
     }
